refactor: create separate chatKeyboard component
