+++ src/frontend/components/ChatKeyboard/style.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+
+export const styles = StyleSheet.create({
+  input: {
+    flex: 1,
+    height: 40,
+    borderWidth: 1,
+    borderRadius: 20,
+    paddingHorizontal: 15
+  },
+});
