+++ src/frontend/screens/ChatUI/style.ts
@@ -22,13 +22,6 @@ export const styles = StyleSheet.create({
     paddingVertical: 5,
     borderTopWidth: 1
   },
-  input: {
-    flex: 1,
-    height: 40,
-    borderWidth: 1,
-    borderRadius: 20,
-    paddingHorizontal: 15
-  },
   centerMessage: {
     flex: 1,
     justifyContent: 'center',
